requirements.txt
absl-py==0.10.0
argon2-cffi==20.1.0
astor==0.8.0
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
conda==4.3.16
cycler==0.10.0
dash==1.4.0
dash-core-components==1.3.0
dash-daq==0.2.1
dash-html-components==1.0.1
dash-renderer==1.1.1
dash-table==4.4.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.0
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.0.1
idna==2.10
importlib-metadata==2.0.0
importlib-resources==1.0.2
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.7
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lightgbm==2.2.3
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mlxtend==0.16.0
mock==3.0.5
mocr==0.3.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
norbert==0.2.1
notebook==6.1.4
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
pandas==0.25.1
pandocfilters==1.4.2
parso==0.7.1
patsy==0.5.1
pickleshare==0.7.5
pid==2.2.3
Pillow==7.2.0
plotly==4.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrealsense2==2.23.0.910
pyrsistent==0.17.3
pyserial==3.4
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2018.3
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.20.2
scipy==1.5.2
seaborn==0.8.1
Send2Trash==1.5.0
Shapely==1.7.1
simple-pid==0.2.1
simplegeneric==0.8.1
six==1.15.0
statsmodels==0.9.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.2.0
xlrd==1.2.0
zipp==3.3.0
